FR 2.1, FR 2.3, FR 2.4, FR 2.5, FR 2.6, FR 2.7
The Resource chart module will be renamed to Resource Calender.

As a user I want to filter the Resource Calendar by single Project so
that I can view my all my Resources for a single project.

As a User I want to to filter on multiple Contact Records on my Resource
Calendar so that I can decide which contacts are to be displayed on the
Resource Calendar.

As a User I want to filter on multiple User Records on my Resource
Calendar so that I can decide which users are to be displayed on the
Resource Calendar.

As a User I want to filter on Project, Multiple Contacts and Multiple
Users so that I can have a granular view of my Projects and Resources.

As a User I want to filter my Resource Calendar by date so that I can
see my resources for a particular time period.