FR 3.9 , 3.10
As a User I want my Project Template Gantt Chart to consider my business
working hours so that the Gantt chart gives a true representation of the
time line.

As a User when in the Edit view of a Project Template record I want to
override the Project Template Gantt chart using the business working
hours so that the Gantt Chart does not consider business working hours.