FR 3.15, 3.16, 3.17, 3.18
As a User when I create a Project from a Project Template I want to be
able to use a 'Copy All Tasks With Resource’s check box to decide if I
want the Project Template Tasks to still have their Resources attached
to them when creating Project from a Project Template.

As a User I want a Help Icon next to the 'Copy All Tasks With Resource's
to explain the purpose of the field.

Given that the 'Copy All Tasks With Resource's check box is not selected
then the 'Copy Tasks with Resource's hidden multi select field will
appear.

As a User when I create a Project from a Project Template I want to be
able to choose from the 'Copy Tasks with Resource's field which tasks
will have their resources copied over from the Project Template to the
Project so that Tasks in the Project can still have their Resource
attached.