FR 3.1, 3.2, 3.3, 3.4
As a User when in the Project Template Edit view I want to click on a
Add Contacts Resource button so that I can add multiple Contact
Resources to my Project Template.

As a User when in the Project Template Edit view I want to click on a
Add User Resource button so that I can add multiple User resources to my
Project Template.

As a User I want to see a Contacts Resource list within my Project
Template Edit view which will display all of the Contact Resources
selected for the Project Template record so that I can see all Contact
Resources selected for the Project Template.

As  a User I want to see a User Resource List within my Project Template
Edit view which will display all of the User resources selected for the
Project Template record so that I can see all User Resources selected
for the Project Template