Changing the behaviour of panels in the detail view.
When action menus are enabled and When there are only panels and there are not any tabs, make the first panel a tab so that the action menu looks correct. This is regardless of what the meta/studio defines the first panel should always be tab.

(cherry picked from commit 8d210a8)