Removing "`" from db query. COLUMN "type" is uneffected.
Quoting the column 'type' because it is a keyword in MySQL is bug.
The responsibility should be in the database manager. I have
removed the quotes and the it seems to work without them.

I can't see why it was added in the first place as both engines on
MySQL and SQL seem to process the query without the need for
quotations or brackets. Perhaps it was just as a precaution.

This change has been tested on:
* Ubuntu 14.04 PHP 5.6 / MySQL 5.6
* Ubuntu 16.04 PHP 5.7 / MySQL 5.7
* Windows Server 2008 R2 / SQL Server 2008 R2

The issue was related to https://github.com/salesagility/SuiteCRM/pull/1525