Fix Issue #2307
Ensure the current active module link is displayed at the bottom of any tab group list even when the module list is greater than the max tabs limit.

As the module list is split into two arrays one for the main module list and one for an extra list which gets pushed to the groupTabs array after the main list in which the active module is contained. Added a condition to check if the list is greater than the max tab limit and if so also push the active module to the extra list also.

To ensure the shortcut tab group is still built correctly for the active module, rather than unsetting the current active tab at this point or simply not pushing it into the main list, it is unset after the array used for the shortcut tab functionality is built by looping through the groupTabs array and checking that the tablist is greater than the tab limit, not in the All tab group and present in both the main and extra list arrays.