Fixed #6485 - Users: Display opt-in ticks also for external email clients
In case the user had selected the external email client all email links would
not display any opt-in information (ticks). This is due to the ticks logic
being implemented in EmailUI which is only used in the sugar client case.
The external client path falls back to a simple mailto link.

This makes things always use EmailUI for rendering email links (including ticks)
and makes EmailUI use simple mailto links in case the user settings are set
to an external client.