Manage subscriptions: Fix unsubscribed users showing up as subscribed sometimes
The code was checking if the user was unsubscribed with the list ID of the
subscription list. This will never work and depending on the iteration order
(if the unsubscribed list gets checked first) this results in the user being
displayed as subscribed even in case he/she is on an exempt list.

This fixes the code to check if the campaign name is in the unsubscribed list
instead, like happens in the case where the non-exempt list comes first.

get_subscription_lists_keyed(), which does the same thing as
get_subscription_lists(), but returns the result in a more structured way,
also uses the campaign name and the same logic for this case.