Fixed #4872 - Fix some async subpanel actions failing if refresh_page=1
In case the subpanel layout uses refresh_page=1 then showSubPanel() will start an
async request and after that reload the page. In many cases this will lead to the
just started request being aborted and the page gets reloaded without anything changing.

This was introduced in 829a06f5b8af where the previously sync call was replaced with an
async one.

This is most visible for me in the "Users" module were I can't remove security groups
through the subpanel and it just refreshes the page without removing the group.

Fix this my moving the page refresh check in the ajax callback so the page gets only reloaded
once the request is finished.