Fix emails losing confirmed opt-in when converting a lead to a contact
When confirmed opt-in is enabled then converting a lead with a confirmed email
address to a contact makes the address lose its confirmation.

AddUpdateEmailAddress only gets a flag if opt-in is set or not, and when trying to update
the existing email record it replaces the COI_STAT_CONFIRMED_OPT_IN state with just
COI_STAT_OPT_IN because of that.

To fix this, in case the passed opt-in flag is 1 and the existing record has some kind
of opt-in state (!COI_STAT_DISABLED) we use that existing state to update the record instead
of forcing COI_STAT_OPT_IN.

In case of a lead conversion where nothing is changed this result in no DB update being
performed on the address.