Fixed #6037 - Reports: Fix adding relations from the first tree node
The main report module is added to the module tree as well, but with an empty
module_path. When selecting related modules the module paths get concatenated
with the root module which is empty in this case and results in a wrong module
path being stored in the report.

This patch works around the issue by checking for an empty module_path before adding
the next module to the path.

Note that this wont fix existing reports that are broken because of this because
the wrong path is saved in the database and persists.

There is really no difference in the children of the first module and the other
toplevel nodes, so ideally the main module wouldn't have any child nodes in the first
place, but this would require more work and can be changed another time.