unittests: Fixes for PHP 7.3
This fixes two types or new errors/warnings in PHP 7.3:

(1) PHP Warning:  "continue" targeting switch is equivalent to "break".
    Did you mean to use "continue 2"?

In all cases the continue was used as a break anyway, so just replace it
with the less confusing break statement.

(2) preg_replace(): Compilation failed: invalid range in character class at offset 4

Fix by properly escaping the "-" in the regex range.

This makes the unit tests pass with PHP 7.3 (the codeception tests still fail)