Fixed #5662 - EmailTemplate: Fix images URLs not being converted with mozaik
addDomainToRelativeImagesSrc() is supposed to convert images elements of the form
"public/UUID.png" to absolute URLs including the site URL. The regex used here
didn't work because the mozaik editor includes a style attribute at the beginning resulting
in nothing getting replaced.

This makes the regex less strict to handle that case and adds some tests while at it.

Fixes #5662