Fix V8 API authorization header passing with apache+php-fpm
With apache+php-fpm the authorization header gets stripped and never
reaches the app. The .htaccess already includes some rules to pass them through
with an env var, but because of the rule order this is never applied.

Fix the rewrite rules by moving the env var rules before the rules marked as [L]ast.

Passing an env var with a rewrite rule results in apache prefixing the env
var with "REDIRECT_". To work around this copy REDIRECT_HTTP_AUTHORIZATION to
HTTP_AUTHORIZATION in case only the former is set.