EmailTemplates: Convert cache/ paths to public/ by default
The functionality was added in 1232e3f81c9959b5c9d but not enabled in
the template save case. Things still ended up in public/ many times
because repairEntryPointImages() handled this if the HTML was in
the right format.

This makes sure that useUploadFolder is always true so that all images
in templates are always moved into public/.