Add a new visitPage helper for tests.
The goal of this helper is to make it easier to navigate through the
site without needing to interact with as much of the site.

If this is used across the entire acceptance test suite, it could save a decent amount of time.

Remove gotoMaps() and replace uses of it with the visitPage helper.