tests: throw an error in case exit() is called during testing
When exit() is called during tests the test suite will stop without returning an error.

To avoid that add our own exit() wrapper which throws an exception instead when
tests are running. There are many more calls to exit() in the code base, this just changes the
common path through sugar_cleanup()